requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
eventlet==0.33.3
python-socketio==5.9.0
python-engineio==4.7.1
gunicorn==21.2.0